refactor: move Glass component to primitives folder
