+++ frontend/src/components/ui/index.ts
@@ -1,12 +1,12 @@
 // Re-export primitives
-export * from './primitives';
+export * from "./primitives";
 
 // Export other UI components
-export { BurgerMenu } from './BurgerMenu';
-export { CampaignFilter } from './CampaignFilter';
-export { default as ImagePlaceholder } from './ImagePlaceholder';
-export { default as Modal } from './Modal';
-export { default as NavigationMenu } from './NavigationMenu';
-export { SidebarMenu } from './SidebarMenu';
-export { default as StatusBadge } from './StatusBadge';
-export { ViewToggle } from './ViewToggle'; 
+export { BurgerMenu } from "./BurgerMenu";
+export { CampaignFilter } from "./CampaignFilter";
+export { default as ImagePlaceholder } from "./ImagePlaceholder";
+export { default as Modal } from "./Modal";
+export { default as NavigationMenu } from "./NavigationMenu";
+export { SidebarMenu } from "./SidebarMenu";
+export { default as StatusBadge } from "./StatusBadge";
+export { ViewToggle } from "./ViewToggle";
