+++ frontend/src/components/ui/primitives/index.ts
@@ -1,8 +1,9 @@
-export { Button } from './Button';
-export { default as Input } from './Input';
-export { default as TextArea } from './TextArea';
-export { ProgressBar } from './ProgressBar';
-export { Badge } from './Badge';
-export { Card } from './Card';
-export { Label } from './Label';
-export { Select } from './Select'; 
+export { Button } from "./Button";
+export { default as Input } from "./Input";
+export { default as TextArea } from "./TextArea";
+export { ProgressBar } from "./ProgressBar";
+export { Badge } from "./Badge";
+export { Card } from "./Card";
+export { Label } from "./Label";
+export { Select } from "./Select";
+export { default as Glass } from "./Glass";
